+++ tests/LineItemTest.php
@@ -7,6 +7,7 @@ use SilverStripe\Dev\SapphireTest;
 use SilverCommerce\OrdersAdmin\Model\LineItem;
 use SilverCommerce\TaxAdmin\Tests\Model\TestProduct;
 use SilverCommerce\CatalogueAdmin\Model\CatalogueProduct;
+use SilverCommerce\OrdersAdmin\Tests\Model\TestProduct;
 
 class LineItemTest extends SapphireTest
 {
@@ -31,21 +32,11 @@ class LineItemTest extends SapphireTest
      *
      * @return void
      */
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
 
         i18n::set_locale('en_GB');
-    }
-
-    /**
-     * Clean up after tear down
-     *
-     * @return void
-     */
-    public function tearDown()
-    {
-        parent::tearDown();
     }
     
     /**
@@ -59,9 +50,9 @@ class LineItemTest extends SapphireTest
         $item_reduced = $this->objFromFixture(LineItem::class, 'reducedtaxitem');
         $item_vat = $this->objFromFixture(LineItem::class, 'taxitemone');
 
-        $this->assertEquals(0, $item_none->TaxPercentage);
-        $this->assertEquals(5, $item_reduced->TaxPercentage);
-        $this->assertEquals(20, $item_vat->TaxPercentage);
+        $this->assertEqualsWithDelta(0, $item_none->TaxPercentage, 0.001);
+        $this->assertEqualsWithDelta(5, $item_reduced->TaxPercentage, 0.001);
+        $this->assertEqualsWithDelta(20, $item_vat->TaxPercentage, 0.001);
     }
 
     /**
@@ -75,9 +66,9 @@ class LineItemTest extends SapphireTest
         $item_reduced = $this->objFromFixture(LineItem::class, 'reducedtaxitem');
         $item_vat = $this->objFromFixture(LineItem::class, 'taxitemone');
 
-        $this->assertEquals(6.50, $item_none->UnitPrice);
-        $this->assertEquals(5.99, $item_reduced->UnitPrice);
-        $this->assertEquals(5.99, $item_vat->UnitPrice);
+        $this->assertEqualsWithDelta(6.50, $item_none->UnitPrice, 0.0001);
+        $this->assertEqualsWithDelta(5.99, $item_reduced->UnitPrice, 0.0001);
+        $this->assertEqualsWithDelta(5.99, $item_vat->UnitPrice, 0.0001);
     }
 
     /**
@@ -91,9 +82,9 @@ class LineItemTest extends SapphireTest
         $item_reduced = $this->objFromFixture(LineItem::class, 'reducedtaxitem');
         $item_vat = $this->objFromFixture(LineItem::class, 'taxitemone');
 
-        $this->assertEquals(0, $item_none->UnitTax);
-        $this->assertEquals(0.2995, $item_reduced->UnitTax);
-        $this->assertEquals(1.198, $item_vat->UnitTax);
+        $this->assertEqualsWithDelta(0, $item_none->UnitTax, 0.01);
+        $this->assertEqualsWithDelta(0.2995, $item_reduced->UnitTax, 0.00001);
+        $this->assertEqualsWithDelta(1.198, $item_vat->UnitTax, 0.0001);
     }
 
     /**
@@ -107,9 +98,9 @@ class LineItemTest extends SapphireTest
         $item_reduced = $this->objFromFixture(LineItem::class, 'reducedtaxitem');
         $item_vat = $this->objFromFixture(LineItem::class, 'taxitemone');
 
-        $this->assertEquals(6.5, $item_none->UnitTotal);
-        $this->assertEquals(6.2895, $item_reduced->UnitTotal);
-        $this->assertEquals(7.188, $item_vat->UnitTotal);
+        $this->assertEqualsWithDelta(6.5, $item_none->UnitTotal, 0.001);
+        $this->assertEqualsWithDelta(6.2895, $item_reduced->UnitTotal, 0.00001);
+        $this->assertEqualsWithDelta(7.188, $item_vat->UnitTotal, 0.0001);
     }
 
     /**
@@ -123,9 +114,9 @@ class LineItemTest extends SapphireTest
         $item_reduced = $this->objFromFixture(LineItem::class, 'reducedtaxitem');
         $item_vat = $this->objFromFixture(LineItem::class, 'taxitemone');
 
-        $this->assertEquals(13.00, $item_none->SubTotal);
-        $this->assertEquals(11.98, $item_reduced->SubTotal);
-        $this->assertEquals(11.98, $item_vat->SubTotal);
+        $this->assertEqualsWithDelta(13.00, $item_none->SubTotal, 0.001);
+        $this->assertEqualsWithDelta(11.98, $item_reduced->SubTotal, 0.001);
+        $this->assertEqualsWithDelta(11.98, $item_vat->SubTotal, 0.001);
     }
 
     /**
@@ -139,9 +130,9 @@ class LineItemTest extends SapphireTest
         $item_reduced = $this->objFromFixture(LineItem::class, 'reducedtaxitem');
         $item_vat = $this->objFromFixture(LineItem::class, 'taxitemone');
 
-        $this->assertEquals(0, $item_none->TaxTotal);
-        $this->assertEquals(0.599, $item_reduced->TaxTotal);
-        $this->assertEquals(2.396, $item_vat->TaxTotal);
+        $this->assertEqualsWithDelta(0, $item_none->TaxTotal, 0.01);
+        $this->assertEqualsWithDelta(0.599, $item_reduced->TaxTotal, 0.0001);
+        $this->assertEqualsWithDelta(2.396, $item_vat->TaxTotal, 0.0001);
     }
 
     /**
@@ -155,9 +146,9 @@ class LineItemTest extends SapphireTest
         $item_reduced = $this->objFromFixture(LineItem::class, 'reducedtaxitem');
         $item_vat = $this->objFromFixture(LineItem::class, 'taxitemone');
 
-        $this->assertEquals(13.00, $item_none->Total);
-        $this->assertEquals(12.579, $item_reduced->Total);
-        $this->assertEquals(14.376, $item_vat->Total);
+        $this->assertEqualsWithDelta(13.00, $item_none->Total, 0.0001);
+        $this->assertEqualsWithDelta(12.579, $item_reduced->Total, 0.0001);
+        $this->assertEqualsWithDelta(14.376, $item_vat->Total, 0.0001);
     }
 
     /**
@@ -227,9 +218,10 @@ class LineItemTest extends SapphireTest
 
         $this->assertTrue(is_object($versioned_product));
         $this->assertEquals("Test versioned product", $versioned_product->Title);
-        $this->assertEquals(1, $versioned_product->Version);
-        $this->assertEquals(4.99, $versioned_product->NoTaxPrice);
-        $this->assertEquals(4.99, $line_item->NoTaxPrice);
+        $this->assertEqualsWithDelta(1, $versioned_product->Version, 0.01);
+        $this->assertEqualsWithDelta(4.99, $versioned_product->NoTaxPrice, 0.001);
+        $this->assertEqualsWithDelta(4.99, $line_item->NoTaxPrice, 0.001);
+        $this->assertEqualsWithDelta(5.99, $product->NoTaxPrice, 0.001);
     }
 
     /**