+++ changelog.md
@@ -231,6 +231,16 @@
 * Ensure that when an estimate is converted to an invoice, it is actually an estimate
 * Forcefully regenerate invoice ref and prefix on conversion (rather than relying on onBeforeWrite)
 
+## 1.4.5
+
+* Add search context from 1.5 branch
+
+## 1.4.6
+
+* Add UUID's to orders
+* Switch to using UUID and Key for frontend display
+* Allow manually setting of an order in OrderFactory
+
 ## 1.5.0
 
 * Add ability to specify if estimate can have a negative value